Fix a bug with touch event handling
