+++ packages/react-resizable-panels-website/src/routes/examples/Collapsible.module.css
@@ -15,7 +15,7 @@
   background-color: var(--color-button-background-hover);
 }
 
-@media (max-width: 500px) {
+@media (pointer: coarse) {
   .ResizeHandle,
   .ResizeHandleCollapsed {
     width: 1rem;
